Regression improvements, psf phot
